+++ src/MemoryDatabase.php
@@ -20,14 +20,20 @@ class MemoryDatabase extends Database
 {
     /**
      * Database constructor.
+     * @param bool $ftsEnabled Whether to enable full text search support (requires FTS5 extension)
+     * @param int $timeout Max time in seconds to obtain a lock
      * @param DatabaseConnection|null $dbConnection
      * @param FileSystemInterface|null $fileSystem
      * @throws DatabaseException
      */
-    public function __construct(?DatabaseConnection $dbConnection = null, ?FileSystemInterface $fileSystem = null)
-    {
+    public function __construct(
+        bool $ftsEnabled = false,
+        int $timeout = 1,
+        ?DatabaseConnection $dbConnection = null,
+        ?FileSystemInterface $fileSystem = null
+    ) {
         $dsn = 'sqlite::memory:';
-        $this->conn = $dbConnection ?? new DatabaseConnection($dsn);
+        $this->conn = $dbConnection ?? new DatabaseConnection($dsn, false, $timeout, $ftsEnabled);
         $this->fileSystem = $fileSystem ?? new FileSystem();
     }
 }